Finish the c++07 under

+++ c++07/ex01/iter.hpp
@@ -6,7 +6,7 @@
 /*   By: aitaouss <aitaouss@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/06 00:48:44 by aitaouss          #+#    #+#             */
-/*   Updated: 2024/10/06 01:04:37 by aitaouss         ###   ########.fr       */
+/*   Updated: 2024/10/07 10:54:17 by aitaouss         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,12 +16,15 @@
 #include <string>
 
 template <typename I, typename F>
-void    iter(I& arr, int len, F funct){
-    for(int i = 0; i < len; ++i){
+void    iter(I& arr, size_t len, F funct)
+{
+    for(size_t i = 0; i < len; ++i)
+    {
         funct(arr[i]);
     }
 }
 
-void    printelement(int i){
+void    printelement(size_t i)
+{
     std::cout << i << std::endl;
 }
